Steer people away from `switchPromptly*`
Create non-prompt counterparts, even though these helper functions are
dramatically simpler, so that no one mistakenly thinks we are endorsing
the prompt versions over their non-prompt equivalents. This is important
because the "prompt" semantics are rarely needed, yet are much more
difficult to use correctly.

Also, rename `switchPromptly` and `switchPromptOnly` to `switchHoldr*`.
This helpers show how they are far more related to each other (having
identical types) than they are to plain `switch`.

One small downside of this change is that it obscures relationship
between `switchPromptlyDyn` and `switchHoldDyn`. This is deemd worth it
as the use of the `current` behavior to avoid holding really does make a
semantic difference: Old inner events prior to the first outer event
firing will never be lost.